Push first version of Python-based flasher
It seems to work - I've successfully flashed the blinky and
rgb images and they both appear to function correctly.

Syntax: ./flasher.py [compressedbistream.bin]

If you don't provide a bitstream file, it will locate and
interrogate the WebFPGA programmer, and do all of the steps
up to but not including actual erasure of the device, and then
stop.